+++ run-docker-dvdrental.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-docker compose -f tests/translation/dvdrental/docker/docker-compose.yml up --wait
-echo "Sleeping for 10 sec before verification.."
-sleep 10
-docker run --network=host --rm rsundqvist/sakila-preload:db-tests