+++ tests/mapping/test_mapper.py
@@ -3,6 +3,7 @@ import pytest
 from rics.mapping import Cardinality, Mapper, exceptions
 from rics.mapping.exceptions import UserMappingError, UserMappingWarning
 from rics.mapping.types import MatchTuple
+from rics.utility.collections.dicts import InheritedKeysDict
 
 
 def _substring_score(k, c, _):
@@ -24,6 +25,7 @@ def test_default(candidates):
 
 def test_with_overrides(candidates):
     mapper: Mapper[str, str, None] = Mapper(overrides={"a": "fixed"})
+    assert not mapper.context_sensitive_overrides
     assert mapper.apply(["a"], candidates).left_to_right == {"a": ("fixed",)}
     assert mapper.apply(["b"], candidates).left_to_right == {"b": ("b",)}
     assert mapper.apply(["a", "b"], candidates).left_to_right == {"a": ("fixed",), "b": ("b",)}
@@ -190,3 +192,40 @@ def test_conflicting_function_overrides_prioritizes_first(values, candidates, ex
     )
 
     assert mapper.apply(values, candidates, override_function=lambda *_: 0).flatten() == expected
+
+
+@pytest.mark.parametrize(
+    "overrides, expected",
+    [
+        (None, False),
+        ({"foo": "bar"}, False),
+        (InheritedKeysDict(), True),
+    ],
+)
+def test_blank_overrides(overrides, expected):
+    assert Mapper(overrides=overrides).context_sensitive_overrides is expected
+
+
+def test_context_sensitive_overrides():
+    mapper = Mapper(
+        overrides=InheritedKeysDict(
+            default={"value0": "default0", "value1": "default1"},
+            specific={
+                0: {"value0": "c0-override-0"},
+                1: {"value0": "c1-override-0"},
+            },
+        )
+    )
+    assert mapper.context_sensitive_overrides
+    values = ["value0", "value1"]
+    assert mapper.apply(values, [], 0).flatten() == {"value0": "c0-override-0", "value1": "default1"}
+    assert mapper.apply(values, [], 1).flatten() == {"value0": "c1-override-0", "value1": "default1"}
+    assert mapper.apply(values, [], 1999).flatten() == {"value0": "default0", "value1": "default1"}
+
+    with pytest.raises(ValueError, match="Must pass a context"):
+        mapper.apply(values, [])
+
+
+def test_copy():
+    assert Mapper() == Mapper()
+    assert Mapper() == Mapper().copy()