+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- The `Mapper.copy` method.
+
+### Changed
+- `Mapper.apply` now require `context` with `InheritedKeysDict`-type overrides.
+
+### Fixed
+- Fixed exact value-candidate match short-circuiting in `HeuristicScore`.
+
+### Removed
+- All `rics.translation` modules. Use [id-translation](https://github.com/rsundqvist/id-translation/) instead.
+
 ## [1.0.1] - 2022-11-26
 
 ### Changed